src/requirements.txt
django
django-mptt
django-mptt-admin
django-taggit
django-csp
lxml

# optional
psycopg2-binary

# developer resources
# django-debug-toolbar
# black
# pyupgrade
